+++ structurizr-core/src/com/structurizr/configuration/Visibility.java
@@ -1,0 +1,8 @@
+package com.structurizr.configuration;
+
+public enum Visibility {
+
+    Private,
+    Public
+
+}